Refactor show/hide legend in plots
 - refactor show/hide legend out of dispatch_plot.py and move it to
   viz/common_functions
 - add show/hide legend functionality to capacity plot

Closes #128
