+++ viz/common_functions.py
@@ -7,6 +7,8 @@
 
 import os.path
 import sqlite3
+from bokeh.models import CustomJS
+from bokeh import events
 
 
 def connect_to_database(parsed_arguments):
@@ -31,3 +33,18 @@ def connect_to_database(parsed_arguments):
     conn = sqlite3.connect(db_path)
 
     return conn
+
+
+def show_hide_legend(plot):
+    """
+    :param plot:
+
+    Show/hide the legend on double tap.
+    """
+    def show_hide_legend_py(legend=plot.legend[0]):
+        legend.visible = not legend.visible
+
+    plot.js_on_event(
+        events.DoubleTap,
+        CustomJS.from_py_func(show_hide_legend_py)
+    )
