+++ viz/dispatch_plot.py
@@ -12,17 +12,16 @@ Make results dispatch plot (by load zone, stage, horizon).
 # TODO: add example df for testing?
 
 from argparse import ArgumentParser
-from bokeh.models import ColumnDataSource, Legend, CustomJS
+from bokeh.models import ColumnDataSource, Legend
 from bokeh.plotting import figure, output_file, show
 from bokeh.models.tools import HoverTool
 from bokeh.embed import json_item
-from bokeh import events
 import pandas as pd
 import os
 import sys
 
 # GridPath modules
-from viz.common_functions import connect_to_database
+from viz.common_functions import connect_to_database, show_hide_legend
 
 
 def parse_arguments(arguments):
@@ -539,22 +538,6 @@ def create_plot(df):
         plot.add_tools(hover)
 
     return plot
-
-
-def show_hide_legend(plot):
-    """
-    :param plot:
-
-    Show/hide the legend on double tap. Something like this can probably be
-    used as a common function to be shared by different plots.
-    """
-    def show_hide_legend_py(legend=plot.legend[0]):
-        legend.visible = not legend.visible
-
-    plot.js_on_event(
-        events.DoubleTap,
-        CustomJS.from_py_func(show_hide_legend_py)
-    )
 
 
 def draw_dispatch_plot(c, scenario_id, load_zone, horizon, stage):
